fastapi_app/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
PyJWT==2.8.0
cryptography==41.0.7
requests==2.31.0

langchain==0.2.16
langchain-chroma==0.1.4
langchain-google-genai==1.0.10
langchain-core==0.2.38
langchain-community==0.2.16
google-generativeai==0.8.3
google-api-core>=2.11.0
chromadb==0.5.6

pypdf==4.3.1
crawl4ai==0.3.74

protobuf==4.25.5
setuptools>=68.0.0
